requirements.txt
Django>=4.2
gunicorn
appwrite
python-decouple
python-dotenv
Pillow
pymongo
gridfs